+++ src/main/java/org/eclipse/keypop/storagecard/UnexpectedCommandStatusException.java
@@ -11,6 +11,7 @@
  ************************************************************************************** */
 package org.eclipse.keypop.storagecard;
 
+import org.eclipse.keypop.reader.selection.InvalidCardResponseException;
 import org.eclipse.keypop.storagecard.card.StorageCard;
 
 /**
@@ -23,7 +24,10 @@ import org.eclipse.keypop.storagecard.card.StorageCard;
  *
  * @since 1.0.0
  */
-public final class UnexpectedCommandStatusException extends StorageCardException {
+public final class UnexpectedCommandStatusException extends InvalidCardResponseException
+    implements StorageCardException {
+
+  private final Integer blockAddress;
 
   /**
    * Creates a new exception indicating a card status code error during the execution of a storage
@@ -35,6 +39,17 @@ public final class UnexpectedCommandStatusException extends StorageCardException
    * @since 1.0.0
    */
   public UnexpectedCommandStatusException(Integer blockAddress, String message, Throwable cause) {
-    super(blockAddress, message, cause);
+    super(message, cause);
+    this.blockAddress = blockAddress;
+  }
+
+  /**
+   * {@inheritDoc}
+   *
+   * @since 1.0.0
+   */
+  @Override
+  public Integer getBlockAddress() {
+    return blockAddress;
   }
 }